Refactors for bot module.
